REFACTORED Code meke one common method for sorting

+++ src/main/java/com/bridgelabz/indiancensusanalyser/model/CensusDAO.java
@@ -14,7 +14,7 @@ public class CensusDAO {
     public double totalArea;
     public int usPopulation;
     public String usState;
-    public int housingUnits;
+    public double housingUnits;
     public double waterArea;
     public String stateId;
